--- packages/rhf-mui-components/package/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "@nish1896/rhf-mui-components",
-  "version": "1.0.2",
-  "description": "Reusable mui components for react-hook-form to minimize your time and effort in creating forms",
-  "author": "Nishant Kohli",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "files": [
-    "dist/",
-    "README.md"
-  ],
-  "scripts": {
-    "build": "tsc --build",
-    "prepublishOnly": "npm run build",
-    "publish:manual": "npm publish --access public",
-    "publish:auto": "npm publish --provenance --access public"
-  },
-	"peerDependencies": {
-		"@mui/icons-material": ">=5",
-    "@mui/material": ">=5",
-    "@mui/x-date-pickers": ">=7",
-		"react-hook-form": ">=7"
-	},
-  "keywords": [
-    "@nish1896",
-    "@nish1896/rhf-mui-components",
-    "react-hook-form",
-    "rhf-mui",
-    "reusable rhf components",
-    "material-ui",
-    "mui",
-    "react form",
-    "@hookform",
-    "@hookform/resolvers",
-    "reusable form components"
-  ],
-  "homepage": "https://github.com/nishkohli96/eslint-config#readme",
-  "repository": {
-    "url": "git+https://github.com/nishkohli96/eslint-config.git"
-  },
-  "license": "ISC"
-}
