--- packages/rhf-mui-components/src/components/checkbox/index.ts
@@ -1,2 +1,0 @@
-export * from './single';
-export * from './group';
