--- packages/rhf-mui-components/src/components/index.ts
@@ -1,7 +1,0 @@
-export * from './checkbox';
-export * from './input';
-export * from './input-controls';
-export * from './miscellaneous';
-export * from './pickers';
-export * from './select';
-
