--- packages/rhf-mui-components/src/components/input-controls/index.ts
@@ -1,4 +1,0 @@
-export * from './radio-group';
-export * from './rating';
-export * from './slider';
-export * from './switch';
