--- packages/rhf-mui-components/src/components/input-controls/rating/index.tsx
@@ -1,80 +1,0 @@
-import { useContext, ReactNode, SyntheticEvent } from 'react';
-import { Control, Controller, Path, FieldValues } from 'react-hook-form';
-import { FormHelperTextProps } from '@mui/material/FormHelperText';
-import { FormLabelProps } from '@mui/material/FormLabel';
-import MuiRating, { RatingProps } from '@mui/material/Rating';
-import { FormControl, FormLabel, FormHelperText } from '../../common';
-import { RHFMuiConfigContext } from '../../../config';
-import { fieldNameToLabel } from '../../../utils';
-
-export type RHFRatingProps<T extends FieldValues> = {
-  fieldName: Path<T>;
-  control: Control<T>;
-  onValueChange?: (
-    e: SyntheticEvent<Element, Event>,
-    newValue: number | null
-  ) => void;
-  label?: ReactNode;
-  showLabelAboveFormField?: boolean;
-  formLabelProps?: Omit<FormLabelProps, 'error'>;
-  helperText?: ReactNode;
-  errorMessage?: ReactNode;
-  hideErrorMessage?: boolean;
-  formHelperTextProps?: Omit<FormHelperTextProps, 'children' | 'error'>;
-} & Omit<RatingProps, 'name' | 'onChange' | 'error' | 'value'>;
-
-export function RHFRating<T extends FieldValues>({
-  fieldName,
-  control,
-  onValueChange,
-  label,
-  showLabelAboveFormField,
-  formLabelProps,
-  helperText,
-  errorMessage,
-  hideErrorMessage,
-  formHelperTextProps,
-  ...rest
-}: RHFRatingProps<T>
-) {
-  const { defaultFormLabelSx, defaultFormHelperTextSx } = useContext(RHFMuiConfigContext);
-  const fieldLabel = label ?? fieldNameToLabel(fieldName);
-  const isError = Boolean(errorMessage);
-
-  return (
-    <Controller
-      name={fieldName}
-      control={control}
-      render={({ field }) => {
-        const { value, onChange, ...otherFieldParams } = field;
-        return (
-          <FormControl error={isError}>
-            <FormLabel
-              label={fieldLabel}
-              isVisible={showLabelAboveFormField}
-              error={isError}
-              formLabelProps={formLabelProps}
-              defaultFormLabelSx={defaultFormLabelSx}
-            />
-            <MuiRating
-              {...rest}
-              {...otherFieldParams}
-              onChange={(e, newValue) => {
-                onChange(Number(newValue));
-                onValueChange && onValueChange(e, newValue);
-              }}
-            />
-            <FormHelperText
-              error={isError}
-              errorMessage={errorMessage}
-              hideErrorMessage={hideErrorMessage}
-              helperText={helperText}
-              defaultFormHelperTextSx={defaultFormHelperTextSx}
-              formHelperTextProps={formHelperTextProps}
-            />
-          </FormControl>
-        );
-      }}
-    />
-  );
-}
