--- packages/rhf-mui-components/src/components/input/index.ts
@@ -1,2 +1,0 @@
-export * from './password';
-export * from './textfield';
