--- packages/rhf-mui-components/src/components/input/textfield/index.tsx
@@ -1,83 +1,0 @@
-import { useContext, ReactNode, ChangeEvent } from 'react';
-import {
-  UseFormRegister,
-  Path,
-  FieldValues,
-  RegisterOptions
-} from 'react-hook-form';
-import { FormHelperTextProps } from '@mui/material/FormHelperText';
-import { FormLabelProps } from '@mui/material/FormLabel';
-import MuiTextField, { TextFieldProps } from '@mui/material/TextField';
-import { FormControl, FormLabel, FormHelperText } from '../../common';
-import { RHFMuiConfigContext } from '../../../config';
-import { fieldNameToLabel } from '../../../utils';
-
-export type RHFTextFieldProps<T extends FieldValues> = {
-  fieldName: Path<T>;
-  register: UseFormRegister<T>;
-  registerOptions?: RegisterOptions<T, Path<T>>;
-  onValueChange?: (
-    e: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
-  ) => void;
-  showLabelAboveFormField?: boolean;
-  formLabelProps?: Omit<FormLabelProps, 'error'>;
-  errorMessage?: ReactNode;
-  hideErrorMessage?: boolean;
-  formHelperTextProps?: Omit<FormHelperTextProps, 'children' | 'error'>;
-} & Omit<TextFieldProps, 'name' | 'onChange' | 'error' | 'value'>;
-
-export function RHFTextField<T extends FieldValues>({
-  fieldName,
-  register,
-  registerOptions,
-  onValueChange,
-  label,
-  showLabelAboveFormField,
-  formLabelProps,
-  helperText,
-  errorMessage,
-  hideErrorMessage,
-  formHelperTextProps,
-  ...rest
-}: RHFTextFieldProps<T>
-) {
-  const { defaultFormLabelSx, defaultFormHelperTextSx } = useContext(RHFMuiConfigContext);
-  const isError = Boolean(errorMessage);
-  const fieldLabel = label ?? fieldNameToLabel(fieldName);
-
-  const { onChange, ...otherRegisterProps } = register(
-    fieldName,
-    registerOptions
-  );
-
-  return (
-    <FormControl error={isError}>
-      <FormLabel
-        label={fieldLabel}
-        isVisible={showLabelAboveFormField}
-        error={isError}
-        formLabelProps={formLabelProps}
-        defaultFormLabelSx={defaultFormLabelSx}
-      />
-      <MuiTextField
-        autoComplete={fieldName}
-        {...rest}
-        onChange={e => {
-          onChange(e);
-          onValueChange && onValueChange(e);
-        }}
-        {...otherRegisterProps}
-        label={!showLabelAboveFormField ? fieldLabel : undefined}
-        error={isError}
-      />
-      <FormHelperText
-        error={isError}
-        errorMessage={errorMessage}
-        hideErrorMessage={hideErrorMessage}
-        helperText={helperText}
-        defaultFormHelperTextSx={defaultFormHelperTextSx}
-        formHelperTextProps={formHelperTextProps}
-      />
-    </FormControl>
-  );
-}
