--- packages/rhf-mui-components/src/components/miscellaneous/index.ts
@@ -1,2 +1,0 @@
-export * from './color-picker';
-export * from './rich-text-editor';
