--- packages/rhf-mui-components/src/components/select/index.ts
@@ -1,2 +1,0 @@
-export * from './mui';
-export * from './native';
