+++ packages/rhf-mui-components/src/mui/index.ts
@@ -1,0 +1,9 @@
+import RHFCheckbox, { RHFCheckboxProps } from './checkbox';
+
+export {
+	RHFCheckbox
+}
+
+export type {
+	RHFCheckboxProps
+}
